wip: refactoring to folders
