+++ main.py
@@ -1,15 +1,8 @@
-users: list[dict] = [
-    {"name": "Dawid", "surname": "Bałuka", "posts": 6000},
-    {"name": "Kewin", "surname": "Czajkowski", "posts": 6002},
-    {"name": "Kamil", "surname": "Gil", "posts": 1_000_000},
-    {"name": "Daniel", "surname": "Błaszczyk", "posts": 6},
-
-]
+from models.data import users
+from utils.crud import show_users
 
 
-def show_users(user_list: list[dict]) -> None:
-    for user in user_list:
-        print(f"Twój znajomy {user['name']} opublikował: {user['posts']} postów")
+
 
 
 if __name__ == "__main__":
