+++ models/data.py
@@ -1,0 +1,7 @@
+users: list[dict] = [
+    {"name": "Dawid", "surname": "Bałuka", "posts": 6000},
+    {"name": "Kewin", "surname": "Czajkowski", "posts": 6002},
+    {"name": "Kamil", "surname": "Gil", "posts": 1_000_000},
+    {"name": "Daniel", "surname": "Błaszczyk", "posts": 6},
+
+]
