+++ utils/crud.py
@@ -1,0 +1,4 @@
+def show_users(user_list: list[dict]) -> None:
+    for user in user_list:
+        print(f"Twój znajomy {user['name']} opublikował: {user['posts']} postów")
+
